% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lava.mixture-package.R
\docType{package}
\name{lava.mixture-package}
\alias{lava.mixture}
\alias{lava.mixture-package}
\title{Estimation and simulation of mixtures of latent variable models.}
\description{
Framwork for estimating parameters from mixtures of
Latent Variable Models. Plugin for the \code{lava} package.
}
\author{
Klaus K. Holst Maintainer: <k.k.holst@biostat.ku.dk>
}
\keyword{package}

