\name{lava.mixture}
\alias{lava.mixture}
\alias{lava.mixture-package}
\docType{package}
\title{
  Estimation and simulation of mixtures of latent variable models
}
\description{
  Framwork for estimating parameters from mixtures of 
  Latent Variable Models. Plugin for the \code{lava} package.
}
\details{
\tabular{ll}{
Package: \tab lava.mixture \cr
Type: \tab Package \cr
Version: \tab 0.2-0 \cr
Date: \tab 2011-01-11 \cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes \cr
}
}
\author{
Klaus K. Holst
Maintainer: <k.k.holst@biostat.ku.dk>
}
\keyword{ package }
